refactor: move plugin implementation to namespaced class
